test(pytest): Use pytest-django
