+++ src/apps/surveys19/tests/test_nosalaryhire.py
@@ -1,5 +1,4 @@
-from django.test import TestCase
-from .setup import setup_fixtures
+from . import TestCase, setup_fixtures
 from apps.surveys19.models import Survey, NoSalaryHire, Month
