+++ src/apps/surveys19/tests/test_shorttermlack.py
@@ -1,12 +1,10 @@
-from django.test import TestCase
-from .setup import setup_fixtures
+from . import TestCase, setup_fixtures
 from apps.surveys19.models import (
     Survey,
     ShortTermLack,
     WorkType,
     Month,
     Product,
-    ProductType,
 )
