+++ src/pytest.ini
@@ -1,0 +1,8 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = config.settings.local
+python_files = test*.py
+addopts = --reuse-db -s
+markers =
+    s18: 2018 Survey test cases.
+    s19: 2019 Survey test cases.
+    s20: 2020 Survey test cases.
